WIP: initial roughed out implementation, not done
