+++ toolsrc/Chloroplast.Tool/Chloroplast.Tool.csproj
@@ -26,5 +26,59 @@
   </ItemGroup>
   <ItemGroup>
     <Folder Include="Commands\" />
+    <Folder Include="ProjectTemplates\" />
+    <Folder Include="ProjectTemplates\conceptual\" />
+  </ItemGroup>
+  <ItemGroup>
+    <None Remove="ProjectTemplates\" />
+    <None Remove="ProjectTemplates\conceptual\" />
+  </ItemGroup>
+  <ItemGroup>
+    <Content Include="..\..\docs\SiteConfig.yml">
+      <Link>ProjectTemplates\conceptual\SiteConfig.yml</Link>
+      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+    </Content>
+    <Content Include="..\..\docs\templates\SiteFrame.cshtml">
+      <Link>ProjectTemplates\conceptual\templates\SiteFrame.cshtml</Link>
+      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+    </Content>
+    <Content Include="..\..\docs\templates\Default.cshtml">
+      <Link>ProjectTemplates\conceptual\templates\Default.cshtml</Link>
+      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+    </Content>
+    <Content Include="..\..\docs\templates\menu.cshtml">
+      <Link>ProjectTemplates\conceptual\templates\menu.cshtml</Link>
+      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+    </Content>
+    <Content Include="..\..\docs\source\Installing\index.md">
+      <Link>ProjectTemplates\conceptual\source\Installing\index.md</Link>
+      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+    </Content>
+  </ItemGroup>
+  <ItemGroup>
+    <None Include="..\..\docs\source\menu.md">
+      <Link>ProjectTemplates\conceptual\source\menu.md</Link>
+      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+    </None>
+    <None Include="..\..\docs\source\index.md">
+      <Link>ProjectTemplates\conceptual\source\index.md</Link>
+      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+    </None>
+    <None Include="..\..\docs\source\cli\index.md">
+      <Link>ProjectTemplates\conceptual\source\cli\index.md</Link>
+      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+    </None>
+    <None Include="..\..\docs\source\templates\index.md">
+      <Link>ProjectTemplates\conceptual\source\templates\index.md</Link>
+      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+    </None>
+    <None Include="..\..\docs\source\assets\main.css">
+      <Link>ProjectTemplates\conceptual\source\assets\main.css</Link>
+      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+    </None>
+    <None Include="..\..\docs\source\apidocs\index.md">
+      <Link>ProjectTemplates\conceptual\source\apidocs\index.md</Link>
+      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+    </None>
   </ItemGroup>
 </Project>
